refactor - move libraries to src folder
