+++ src/executors/query_validator.py
